split contexts/accounts
