+++ js/packages/common/src/actions/auction.ts
@@ -8,7 +8,7 @@ import {
 import { programIds } from '../utils/programIds';
 import { deserializeUnchecked, serialize } from 'borsh';
 import BN from 'bn.js';
-import { AccountParser } from '../contexts';
+import { AccountParser } from '../contexts/accounts/types';
 import moment from 'moment';
 import { findProgramAddress, StringPublicKey, toPublicKey } from '../utils';
 export const AUCTION_PREFIX = 'auction';
