+++ js/packages/common/src/contexts/accounts/accounts.tsx
@@ -1,0 +1,286 @@
+import React, { useCallback, useContext, useEffect, useState } from 'react';
+import { useWallet } from '@solana/wallet-adapter-react';
+import { AccountInfo, Connection, PublicKey } from '@solana/web3.js';
+import { AccountLayout, MintInfo, u64 } from '@solana/spl-token';
+import { useConnection } from '../../contexts/connection';
+import { TokenAccount } from '../../models';
+import { StringPublicKey, WRAPPED_SOL_MINT } from '../../utils/ids';
+import { programIds } from '../../utils/programIds';
+import { genericCache, cache } from './cache';
+import { deserializeAccount } from './deserialize';
+import { TokenAccountParser, MintParser } from './parsesrs';
+
+const AccountsContext = React.createContext<any>(null);
+
+export const useAccountsContext = () => {
+  const context = useContext(AccountsContext);
+
+  return context;
+};
+
+function wrapNativeAccount(
+  pubkey: StringPublicKey,
+  account?: AccountInfo<Buffer>,
+): TokenAccount | undefined {
+  if (!account) {
+    return undefined;
+  }
+
+  const key = new PublicKey(pubkey);
+
+  return {
+    pubkey: pubkey,
+    account,
+    info: {
+      address: key,
+      mint: WRAPPED_SOL_MINT,
+      owner: key,
+      amount: new u64(account.lamports),
+      delegate: null,
+      delegatedAmount: new u64(0),
+      isInitialized: true,
+      isFrozen: false,
+      isNative: true,
+      rentExemptReserve: null,
+      closeAuthority: null,
+    },
+  };
+}
+
+const UseNativeAccount = () => {
+  const connection = useConnection();
+  const { publicKey } = useWallet();
+
+  const [nativeAccount, setNativeAccount] = useState<AccountInfo<Buffer>>();
+
+  const updateCache = useCallback(
+    account => {
+      if (publicKey) {
+        const wrapped = wrapNativeAccount(publicKey.toBase58(), account);
+        if (wrapped !== undefined) {
+          const id = publicKey.toBase58();
+          cache.registerParser(id, TokenAccountParser);
+          genericCache.set(id, wrapped as TokenAccount);
+          cache.emitter.raiseCacheUpdated(id, false, TokenAccountParser, true);
+        }
+      }
+    },
+    [publicKey],
+  );
+
+  useEffect(() => {
+    let subId = 0;
+    const updateAccount = (account: AccountInfo<Buffer> | null) => {
+      if (account) {
+        updateCache(account);
+        setNativeAccount(account);
+      }
+    };
+
+    (async () => {
+      if (!connection || !publicKey) {
+        return;
+      }
+
+      const account = await connection.getAccountInfo(publicKey);
+      updateAccount(account);
+
+      subId = connection.onAccountChange(publicKey, updateAccount);
+    })();
+
+    return () => {
+      if (subId) {
+        connection.removeAccountChangeListener(subId);
+      }
+    };
+  }, [setNativeAccount, publicKey, connection, updateCache]);
+
+  return { nativeAccount };
+};
+
+const PRECACHED_OWNERS = new Set<string>();
+const precacheUserTokenAccounts = async (
+  connection: Connection,
+  owner?: PublicKey,
+) => {
+  if (!owner) {
+    return;
+  }
+
+  // used for filtering account updates over websocket
+  PRECACHED_OWNERS.add(owner.toBase58());
+
+  // user accounts are updated via ws subscription
+  const accounts = await connection.getTokenAccountsByOwner(owner, {
+    programId: programIds().token,
+  });
+
+  accounts.value.forEach(info => {
+    cache.add(info.pubkey.toBase58(), info.account, TokenAccountParser);
+  });
+};
+
+export function AccountsProvider({ children = null as any }) {
+  const connection = useConnection();
+  const { publicKey } = useWallet();
+  const [tokenAccounts, setTokenAccounts] = useState<TokenAccount[]>([]);
+  const [userAccounts, setUserAccounts] = useState<TokenAccount[]>([]);
+  const { nativeAccount } = UseNativeAccount();
+  const walletKey = publicKey?.toBase58();
+
+  const selectUserAccounts = useCallback(() => {
+    return cache
+      .byParser(TokenAccountParser)
+      .map(id => cache.get(id))
+      .filter(a => a && a.info.owner.toBase58() === walletKey)
+      .map(a => a as TokenAccount);
+  }, [walletKey]);
+
+  useEffect(() => {
+    const accounts = selectUserAccounts().filter(
+      a => a !== undefined,
+    ) as TokenAccount[];
+    setUserAccounts(accounts);
+  }, [nativeAccount, tokenAccounts, selectUserAccounts]);
+
+  useEffect(() => {
+    const subs: number[] = [];
+    cache.emitter.onCache(args => {
+      if (args.isNew && args.isActive) {
+        let id = args.id;
+        let deserialize = args.parser;
+        connection.onAccountChange(new PublicKey(id), info => {
+          cache.add(id, info, deserialize);
+        });
+      }
+    });
+
+    return () => {
+      subs.forEach(id => connection.removeAccountChangeListener(id));
+    };
+  }, [connection]);
+
+  useEffect(() => {
+    if (!connection || !publicKey) {
+      setTokenAccounts([]);
+    } else {
+      precacheUserTokenAccounts(connection, publicKey).then(() => {
+        setTokenAccounts(selectUserAccounts());
+      });
+
+      // This can return different types of accounts: token-account, mint, multisig
+      // TODO: web3.js expose ability to filter.
+      // this should use only filter syntax to only get accounts that are owned by user
+      const tokenSubID = connection.onProgramAccountChange(
+        programIds().token,
+        info => {
+          // TODO: fix type in web3.js
+          const id = info.accountId as unknown as string;
+          // TODO: do we need a better way to identify layout (maybe a enum identifing type?)
+          if (info.accountInfo.data.length === AccountLayout.span) {
+            const data = deserializeAccount(info.accountInfo.data);
+
+            if (PRECACHED_OWNERS.has(data.owner.toBase58())) {
+              cache.add(id, info.accountInfo, TokenAccountParser);
+              setTokenAccounts(selectUserAccounts());
+            }
+          }
+        },
+        'singleGossip',
+      );
+
+      return () => {
+        connection.removeProgramAccountChangeListener(tokenSubID);
+      };
+    }
+  }, [connection, publicKey, selectUserAccounts]);
+
+  return (
+    <AccountsContext.Provider
+      value={{
+        userAccounts,
+        nativeAccount,
+      }}
+    >
+      {children}
+    </AccountsContext.Provider>
+  );
+}
+
+export function useNativeAccount() {
+  const context = useContext(AccountsContext);
+  return {
+    account: context.nativeAccount as AccountInfo<Buffer>,
+  };
+}
+
+export function useMint(key?: string | PublicKey) {
+  const connection = useConnection();
+  const [mint, setMint] = useState<MintInfo>();
+
+  const id = typeof key === 'string' ? key : key?.toBase58();
+
+  useEffect(() => {
+    if (!id) {
+      return;
+    }
+
+    cache
+      .query(connection, id, MintParser)
+      .then(acc => setMint(acc.info as any))
+      .catch(err => console.log(err));
+
+    const dispose = cache.emitter.onCache(e => {
+      const event = e;
+      if (event.id === id) {
+        cache
+          .query(connection, id, MintParser)
+          .then(mint => setMint(mint.info as any));
+      }
+    });
+    return () => {
+      dispose();
+    };
+  }, [connection, id]);
+
+  return mint;
+}
+
+export function useAccount(pubKey?: PublicKey) {
+  const connection = useConnection();
+  const [account, setAccount] = useState<TokenAccount>();
+
+  const key = pubKey?.toBase58();
+  useEffect(() => {
+    const query = async () => {
+      try {
+        if (!key) {
+          return;
+        }
+
+        const acc = await cache
+          .query(connection, key, TokenAccountParser)
+          .catch(err => console.log(err));
+        if (acc) {
+          setAccount(acc);
+        }
+      } catch (err) {
+        console.error(err);
+      }
+    };
+
+    query();
+
+    const dispose = cache.emitter.onCache(e => {
+      const event = e;
+      if (event.id === key) {
+        query();
+      }
+    });
+    return () => {
+      dispose();
+    };
+  }, [connection, key]);
+
+  return account;
+}
