+++ js/packages/common/src/contexts/accounts/parsesrs.ts
@@ -1,0 +1,64 @@
+import { AccountInfo } from '@solana/web3.js';
+import { TokenAccount } from '../../models';
+import { ParsedAccountBase } from './types';
+import { deserializeMint, deserializeAccount } from './deserialize';
+import { StringPublicKey } from '../../utils';
+
+export const MintParser = (
+  pubKey: StringPublicKey,
+  info: AccountInfo<Buffer>,
+) => {
+  const buffer = Buffer.from(info.data);
+
+  const data = deserializeMint(buffer);
+
+  const details = {
+    pubkey: pubKey,
+    account: {
+      ...info,
+    },
+    info: data,
+  } as ParsedAccountBase;
+
+  return details;
+};
+
+export const TokenAccountParser = (
+  pubKey: StringPublicKey,
+  info: AccountInfo<Buffer>,
+) => {
+  // Sometimes a wrapped sol account gets closed, goes to 0 length,
+  // triggers an update over wss which triggers this guy to get called
+  // since your UI already logged that pubkey as a token account. Check for length.
+  if (info.data.length > 0) {
+    const buffer = Buffer.from(info.data);
+    const data = deserializeAccount(buffer);
+
+    const details = {
+      pubkey: pubKey,
+      account: {
+        ...info,
+      },
+      info: data,
+    } as TokenAccount;
+
+    return details;
+  }
+};
+
+export const GenericAccountParser = (
+  pubKey: StringPublicKey,
+  info: AccountInfo<Buffer>,
+) => {
+  const buffer = Buffer.from(info.data);
+
+  const details = {
+    pubkey: pubKey,
+    account: {
+      ...info,
+    },
+    info: buffer,
+  } as ParsedAccountBase;
+
+  return details;
+};
