+++ js/packages/common/src/contexts/accounts/types.ts
@@ -1,0 +1,17 @@
+import { AccountInfo } from '@solana/web3.js';
+import { StringPublicKey } from '../../utils';
+
+export interface ParsedAccountBase {
+  pubkey: StringPublicKey;
+  account: AccountInfo<Buffer>;
+  info: any; // TODO: change to unknown
+}
+
+export type AccountParser = (
+  pubkey: StringPublicKey,
+  data: AccountInfo<Buffer>,
+) => ParsedAccountBase | undefined;
+
+export interface ParsedAccount<T> extends ParsedAccountBase {
+  info: T;
+}
