+++ js/packages/web/src/contexts/index.tsx
@@ -1,2 +1,2 @@
-export * from '@oyster/common/dist/lib/contexts/meta/meta'
+export * from '@oyster/common/dist/lib/contexts/meta/meta';
 export * from './coingecko';
