+++ js/packages/web/src/views/admin/index.tsx
@@ -12,7 +12,10 @@ import {
   Divider,
 } from 'antd';
 import { useMeta } from '../../contexts';
-import { Store, WhitelistedCreator } from '@oyster/common/dist/lib/models/metaplex/index';
+import {
+  Store,
+  WhitelistedCreator,
+} from '@oyster/common/dist/lib/models/metaplex/index';
 import {
   MasterEditionV1,
   notify,
